#7 - Reorganized atmospheres into their own static classes just for cleanliness
